Initial search path now defaults to empty string ''.
Previously connection couldn't estabilished, because right at the start
of building it failed with an Exception. We want the ability to connect to the
database initially, but not QUERY data from the incorrect tables.
So, when building the connection, the client will be able to make QUERIES
explicitly, but need to set the search_path to query as usual.